isMyCommand feature in CommandBus
When I add a bot as an admin with private mode turned off, it gets all the chat messages and all the commands, including commands from other bots.

Since the name of the bot is in the format / command @ MyBot, it is simply discarded in the CommandBase :: parseCommand (telegram-bot-sdk / src / Commands / CommandBus.pxp: 156) method, there is no regular possibility inside the command to verify that this command is addressed to my bot.

A possible workaround is to check the update-> message-> text for the fact that it contains the name of the bot, but several commands can be passed in one line to different bots.

The first approximation of a permanent solution in the proposed patch.
If it is suitable - I will write tests for it, if not, please indicate what to modify.