Bot API 4.2 April 14, 2019
Added support for native polls: added the object Poll, the methods sendPoll and stopPoll and the field poll in the Message and Update objects.
The method deleteMessage can now be used to delete messages sent by a user to the bot in private chats within 48 hours.
Added support for pinned messages in basic groups in addition to supergroups and channel chats: you can pass group's chat_id to pinChatMessage and unpinChatMessage, and receive the pinned group message in Chat object.
Added the field is_member to the ChatMember object, which can be used to find whether a restricted user is a member of the chat.
Added the field forward_sender_name to the Message object, containing name of the sender who has opted to hide their account.