Bot API 4.3 May 31, 2019
Added support for Seamless Telegram Login on external websites (needs Telegram 5.7 or higher).
Added the new object LoginUrl and the new field login_url to the InlineKeyboardButton object which allows to automatically authorize users before they go to a URL specified by the bot. Users will be asked to confirm authorization in their Telegram app (needs version 5.7 or higher) when they press the button:
Added the field reply_markup to the Message object, containing the inline keyboard attached to the message.