Bot API 4.7
March 30, 2020

Added the method sendDice for sending a dice message, which will have a random value from 1 to 6. (Yes, we're aware of the “proper” singular of die. But it's awkward, and we decided to help it change. One dice at a time!)
Added the field dice to the Message object.
Added the method getMyCommands for getting the current list of the bot's commands.
Added the method setMyCommands for changing the list of the bot's commands through the Bot API instead of @BotFather.
Added the ability to create animated sticker sets by specifying the parameter tgs_sticker instead of png_sticker in the method createNewStickerSet.
Added the ability to add animated stickers to sets created by the bot by specifying the parameter tgs_sticker instead of png_sticker in the method addStickerToSet.
Added the field thumb to the StickerSet object.
Added the ability to change thumbnails of sticker sets created by the bot using the method setStickerSetThumb.

Please note. Testing the stickerset thumbnails was not done. Don't have the time to be making stickers!