Allow to add commands stored in an iterable
Adding several commands with a single method call can be really useful,
especially when working with a framework that can detect all commands
instances and inject them in the command bus automatically.

This commit allows commands registration using an iterable, enhancing
the possibilities to — for instance — lazy-load the commands only when
the SDK API is actually called.

Below is an example of how the `Api` class is declared in a recent
Symfony application:

```php
return static function (ContainerConfigurator $container): void {
    $container
        ->services()

        // Tagging all commands, so they can be retrieved below
        ->instanceof(\Telegram\Bot\Commands\CommandInterface::class)
        ->tag('telegram.command')

        // Registering the `Api` class
        ->set(Api::class)
        ->args([
            '$token' => env('TELEGRAM_BOT_TOKEN')->resolve(),
        ])
        // Adding all commands that were previously tagged
        ->call('addCommands', [tagged_iterator('telegram.command')]);
};
```

The example above would not work without changing the type of the
parameter, because the `tagged_iterator` is not an array.